# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: goals.ac API — Programmatic SEO Growth Roadmap Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /geo-audits:
    post:
      operationId: createGeoAudit
      tags: [seo-tools]
      summary: Run a GEO audit on a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeoAuditRequest"
      responses:
        "201":
          description: GEO audit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoAudit"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: URL fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /geo-audits/{id}:
    get:
      operationId: getGeoAudit
      tags: [seo-tools]
      summary: Get a GEO audit by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GEO audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoAudit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GeoIssue:
      type: object
      properties:
        check:
          type: string
        status:
          type: string
          enum: [pass, fail, warn]
        detail:
          type: string
        fix:
          type: string
      required: [check, status, detail, fix]
    CreateGeoAuditRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        roadmap_id:
          type: integer
        website_project_id:
          type: integer
      required: [url]
    GeoAudit:
      type: object
      properties:
        id:
          type: integer
        roadmapId:
          type: integer
          nullable: true
        websiteProjectId:
          type: integer
          nullable: true
        url:
          type: string
        geoScore:
          type: integer
        issues:
          type: array
          items:
            $ref: "#/components/schemas/GeoIssue"
        pageTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        hasSchemaOrg:
          type: boolean
        schemaTypes:
          type: array
          items:
            type: string
        h1Count:
          type: integer
        imageCount:
          type: integer
        imagesMissingAlt:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - geoScore
        - issues
        - hasSchemaOrg
        - schemaTypes
        - h1Count
        - imageCount
        - imagesMissingAlt
        - createdAt
...